Version 0.0.49
(1) Implemented a feature to ignore directories.

(2) Adjust sort logic to place locals block first.

(3) Within any given Terraform file the locals block provides information
that is likely very useful to know in the context of reading other
blocks and as such placing the locals block first in the file provides a
useful ordering.

(4) Changed "source" to only be a meta argument for modules